+++ package.json
@@ -10,6 +10,7 @@
     "axios": "^1.5.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-icons": "^4.11.0",
     "react-redux": "^8.1.3",
     "react-router": "^6.16.0",
     "react-router-dom": "^6.16.0",
