+++ src/components/Header.jsx
@@ -1,24 +1,36 @@
-import { NavLink, Outlet } from 'react-router-dom';
+import { NavLink, Outlet, useLocation } from 'react-router-dom';
+import { AiOutlineHome, AiFillSetting } from 'react-icons/ai';
+import { FaSearchLocation } from 'react-icons/fa';
 
-const Header = () => (
-  <>
-    <header className="header">
-      <nav className="navbar">
-        <ul className="nav-menu">
-          <li className="nav-item">
-            <NavLink to="/" className="navlink">home</NavLink>
-          </li>
-          <li className="nav-item">
-            Home Page
-          </li>
-          <li className="nav-item">
-            <NavLink to="/details" className="navlink">details</NavLink>
-          </li>
-        </ul>
-      </nav>
-    </header>
-    <Outlet />
-  </>
-);
+const Header = () => {
+  const location = useLocation();
+  return (
+    <>
+      <header className="header">
+        <nav className="navbar">
+          <ul className="nav-menu">
+            <li className="nav-item">
+              <NavLink to="/" className="navlink">
+                <AiOutlineHome />
+              </NavLink>
+            </li>
+            <li className="nav-item">
+              <p>{location.pathname === '/' ? 'HOME' : 'DETAILS'}</p>
+            </li>
+            <li className="nav-item">
+              <NavLink to="" className="navlink">
+                <FaSearchLocation />
+              </NavLink>
+            </li>
+            <li className="nav-item">
+              <AiFillSetting />
+            </li>
+          </ul>
+        </nav>
+      </header>
+      <Outlet />
+    </>
+  );
+};
 
 export default Header;
